pyproject.toml
[tool.poetry]
name = "pju"
version = "0.1.0"
description = ""
authors = ["Andraz Vrhovec <andraz@vrhovec.si>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
urllib3 = "^1.26.14"
pandas = "^1.5.3"
frictionless = "^5.6.3"


[tool.poetry.group.dev.dependencies]
vcrpy = "^4.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
